Refactored EnrichedSpanUtils to EnrichedSpanHeaderUtils and EnrichedSpanCookieUtils
